controller/requirements.txt
# HAS Controller required Python packages
circuits
paho-mqtt
requests
PyYAML
websocket-client
